combine both credentials api test files into one
